refactor(InputParser): 서비스 로직에서 파싱 로직 분리
